requirements.txt.txt
streamlit==0.79.0
pandas==1.2.3
numpy==1.18.5
paho-mqtt==1.6.1
python-time==0.3.0